web - scaffold a basic dev setup via "yarn web"
