+++ src/vs/code/browser/workbench/workbench.html
@@ -1,0 +1,12 @@
+<!-- Copyright (C) Microsoft Corporation. All rights reserved. -->
+<!DOCTYPE html>
+<html>
+	<head>
+		<meta charset="utf-8" />
+	</head>
+	<body class="vs-dark" aria-label="">
+	</body>
+
+	<!-- Startup via workbench.js -->
+	<script src="../../../../../out/vs/code/browser/workbench/workbench.js"></script>
+</html>
