--- src/vs/code/electron-browser/workbench/workbench.nodeless.html
@@ -1,12 +1,0 @@
-<!-- Copyright (C) Microsoft Corporation. All rights reserved. -->
-<!DOCTYPE html>
-<html>
-	<head>
-		<meta charset="utf-8" />
-	</head>
-	<body class="monaco-shell vs-dark" aria-label="">
-	</body>
-
-	<!-- Startup via workbench.nodeless.js -->
-	<script src="workbench.nodeless.js"></script>
-</html>
