+++ src/vs/platform/environment/common/environment.ts
@@ -69,7 +69,6 @@ export interface ParsedArgs {
 	'driver'?: string;
 	'driver-verbose'?: boolean;
 	remote?: string;
-	'nodeless'?: boolean; // TODO@ben revisit electron5 nodeless support
 }
 
 export const IEnvironmentService = createDecorator<IEnvironmentService>('environmentService');
