+++ src/vs/workbench/workbench.web.main.css
