+++ src/vs/workbench/workbench.web.main.nls.js
