+++ src/vs/workbench/workbench.web.main.ts
@@ -12,7 +12,7 @@ import 'vs/editor/editor.all';
 // import 'vs/workbench/electron-browser/main.contribution';
 import 'vs/workbench/browser/workbench.contribution';
 
-import 'vs/workbench/browser/nodeless.main';
+import 'vs/workbench/browser/web.main';
 
 //#endregion
 
@@ -94,7 +94,7 @@ import { IBroadcastService, NullBroadcastService } from 'vs/workbench/services/b
 import { ConfigurationResolverService } from 'vs/workbench/services/configurationResolver/browser/configurationResolverService';
 import { IConfigurationResolverService } from 'vs/workbench/services/configurationResolver/common/configurationResolver';
 
-import 'vs/workbench/browser/nodeless.simpleservices';
+import 'vs/workbench/browser/web.simpleservices';
 import 'vs/platform/dialogs/browser/dialogService';
